refactor: 동적 라우팅 변경 전 폴더 및 파일 제거
