--- src/app/[userId]/setting/asdf/page.tsx
@@ -1,57 +1,0 @@
-"use client";
-
-import React from "react";
-import SettingPrivacy from "../components/SettingPrivacy";
-import { getSetting } from "../server-action/settingAction";
-import { useQuery } from "@tanstack/react-query";
-import { Setting } from "@/types/setting";
-import queryKey from "@/queries/queryKey";
-import SettingShowList from "../components/SettingShowList";
-import SettingTabList from "../components/SettingTabList";
-import { useUserInfoStore } from "@/store/useUserInfoStore";
-import { getLoginUserId } from "@/utils/supabase/user";
-
-const testUserId = "588a4dea-b95a-4836-b6bc-10dbafa4a81f";
-
-type Props = {
-  params: {
-    userId: string;
-  };
-};
-
-const page = ({ params }: Props) => {
-  // const { id: userId } = useUserInfoStore();
-
-  const { data: userId } = useQuery<string | undefined>({
-    queryKey: queryKey.auth.loginUser,
-    queryFn: () => getLoginUserId()
-  });
-  const { userId: hostUserId } = params;
-
-  console.log("userId", userId);
-  console.log("hostUserId", hostUserId);
-
-  const { data: setting } = useQuery<Setting>({
-    queryKey: queryKey.setting.setting,
-    queryFn: () => getSetting(testUserId),
-    enabled: !!userId
-  });
-
-  if (!setting) {
-    return <></>;
-  }
-  //TODO: 조건 하나 더 추가 해야함 (params랑 로그인 userId 다른 다른 페이지로)
-  if (userId !== hostUserId) {
-    return <>잉</>;
-  }
-
-  return (
-    <div>
-      <SettingPrivacy setting={setting}></SettingPrivacy>
-      <SettingShowList setting={setting}></SettingShowList>
-      <SettingTabList setting={setting}></SettingTabList>
-    </div>
-  );
-};
-
-export default page;
