--- src/app/setting/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const page = () => {
-  return (
-    <div>ASDFMKLASDFKML</div>
-  )
-}
-
-export default page
